+++ ocplib/ocplib.h
@@ -30,10 +30,10 @@ void filterCutoutOMP ( uint32_t *, int , uint32_t *, int );
 void filterCutoutOMPCache ( uint32_t *, int, uint32_t *, int );
 
 // Declaring the annotateCube implementation
-int annotateCube ( uint32_t * , int , int * , int , int * , uint32_t [][3] , int , char, uint32_t [][3] );
+int annotateCube ( uint32_t * , int , int * , int , uint32_t * , uint32_t [][3] , int , char, uint32_t [][3] );
 
 // Declaring the locateCube implementation
-void locateCube ( uint32_t [][4] , int , uint32_t [][3] , int , int * );
+void locateCube ( uint64_t [][4] , int , uint32_t [][3] , int , int * );
 
 // Declaring XYZMorton zindex function
 int XYZMorton ( uint64_t * );
@@ -45,4 +45,4 @@ void MortonXYZ ( uint64_t , uint64_t [3] );
 void recolorSlice ( uint32_t * , int , int , uint32_t * , uint32_t * ); 
 
 // Declaring Quick Sort function
-void quicksort ( uint32_t [][4] , int ); 
+void quicksort ( uint64_t [][4] , int );
